feat: add date range filter for leads
The changes in this commit add a date range filter for leads in the client application. The filter allows users to specify a start and end date to narrow down the list of leads displayed. This feature improves the usability of the application by providing a way to view leads within a specific time frame.